move servicehook model to sentry_apps
