+++ src/sentry/models/__init__.py
@@ -107,7 +107,6 @@ from .savedsearch import *  # NOQA
 from .scheduledeletion import *  # NOQA
 from .search_common import *  # NOQA
 from .sentryshot import *  # NOQA
-from .servicehook import *  # NOQA
 from .sourcemapprocessingissue import *  # NOQA
 from .statistical_detectors import *  # NOQA
 from .team import *  # NOQA
