+++ src/sentry/sentry_apps/models/__init__.py
@@ -1,4 +1,5 @@
 from .sentry_app import SentryApp
 from .sentry_app_installation import SentryAppInstallation
+from .servicehook import ServiceHook
 
-__all__ = ("SentryApp", "SentryAppInstallation")
+__all__ = ("SentryApp", "SentryAppInstallation", "ServiceHook")
