+++ src/sentry/sentry_apps/services/hook/serial.py
@@ -1,4 +1,4 @@
-from sentry.models.servicehook import ServiceHook
+from sentry.sentry_apps.models.servicehook import ServiceHook
 from sentry.sentry_apps.services.hook import RpcServiceHook
