+++ src/sentry/tasks/servicehooks.py
@@ -2,7 +2,7 @@ from time import time
 
 from sentry.api.serializers import serialize
 from sentry.http import safe_urlopen
-from sentry.models.servicehook import ServiceHook
+from sentry.sentry_apps.models.servicehook import ServiceHook
 from sentry.silo.base import SiloMode
 from sentry.tasks.base import instrumented_task, retry
 from sentry.tsdb.base import TSDBModel
