+++ tests/sentry/api/endpoints/test_project_servicehook_stats.py
@@ -1,5 +1,5 @@
 from sentry import tsdb
-from sentry.models.servicehook import ServiceHook
+from sentry.sentry_apps.models.servicehook import ServiceHook
 from sentry.testutils.cases import APITestCase
 from sentry.tsdb.base import TSDBModel
