+++ tests/sentry/deletions/test_project.py
@@ -18,7 +18,6 @@ from sentry.models.release import Release
 from sentry.models.releasecommit import ReleaseCommit
 from sentry.models.repository import Repository
 from sentry.models.rulesnooze import RuleSnooze
-from sentry.models.servicehook import ServiceHook
 from sentry.monitors.models import (
     CheckInStatus,
     Monitor,
@@ -27,6 +26,7 @@ from sentry.monitors.models import (
     MonitorType,
     ScheduleType,
 )
+from sentry.sentry_apps.models.servicehook import ServiceHook
 from sentry.snuba.models import QuerySubscription, SnubaQuery
 from sentry.testutils.cases import APITestCase, TransactionTestCase
 from sentry.testutils.helpers.datetime import before_now, iso_format
